update basic login and mfa login in sample
