+++ samples/templates/express-direct-auth/web-server/routes/index.js
@@ -1,4 +1,6 @@
 const login = require('./login');
+const basicLogin = require('./basic-login');
+const multifactorLogin = require('./multifactor-login');
 const logout = require('./logout');
 const signup = require('./signup');
 const recoverPassword = require('./recover-password');
@@ -8,6 +10,8 @@ const protected = require('./protected');
 const terminal = require('./terminal');
 
 module.exports = [
+  basicLogin,
+  multifactorLogin,
   login,
   logout,
   signup,
