+++ samples/templates/express-direct-auth/web-server/routes/login/index.js
@@ -1,8 +1,6 @@
 const express = require('express');
-const basic = require('./basic');
 const withWidget = require('./with-widget');
 const withIdp = require('./with-idp');
-const withAuthenticators = require('./with-authenticators');
 const { 
   getAuthClient,
 } = require('../../utils');
@@ -10,10 +8,8 @@ const {
 const router = express.Router();
 
 router.use('/login', [
-  basic,
   withWidget,
   withIdp,
-  withAuthenticators,
 ]);
 
 router.get('/login/callback', async (req, res) => {
