+++ samples/templates/express-direct-auth/web-server/views/error.mustache
@@ -1,9 +1,24 @@
 <!DOCTYPE html>
-<html lang="en">
+<html>
   <head>
     {{>head}}
   </head>
   <body id="samples">
-    {{>errors}}
+  
+    {{>menu}}
+
+    <div class="okta-sign-in-form ui center aligned relaxed grid">
+      <div class="row">
+        <div class="ui header">
+          <h2>Error</h2>
+        </div>
+      </div>
+
+      <div class="row">
+        {{>errors}}
+      </div>
+    </div>
+    
   </body>
 </html>
+
