+++ samples/templates/express-direct-auth/web-server/views/home.mustache
@@ -39,7 +39,7 @@
             <div>
             <ul>
               <li><a id="authenticate-password" href="/login/basic">Login with Username and Password</a></li>
-              <li><a href="/login/with-authenticators">Login with authenticators</a></li>
+              <li><a href="/multifactor-login">Login with Password and Another Factor</a></li>
               <li><a href="/login/with-widget">Self hosted widget</a></li>
               <li><a href="/login/with-idp">Login using an external identity provider</a></li>
               <li><a href="/signup">Self service registration</a></li>
